# packages/books-service/src/schema.swagger.json
{
  "openapi": "3.0.0",
  "paths": {
    "/api": {
      "get": {
        "operationId": "AppController_getHello",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        }
      }
    },
    "/api/v1/book/{id}": {
      "get": {
        "operationId": "BooksController_get",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "default": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Book"
                }
              }
            }
          }
        },
        "tags": [
          "book"
        ]
      }
    },
    "/api/v1/book": {
      "get": {
        "operationId": "BooksController_getList",
        "parameters": [
          {
            "name": "keyword",
            "required": false,
            "in": "query",
            "schema": {
              "nullable": true,
              "type": "string"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "nullable": true,
              "default": 1,
              "type": "number"
            }
          },
          {
            "name": "size",
            "required": false,
            "in": "query",
            "description": "page size",
            "schema": {
              "nullable": true,
              "default": 10,
              "type": "number"
            }
          },
          {
            "name": "start",
            "required": false,
            "in": "query",
            "description": "parse by new Date()",
            "schema": {
              "format": "date-time",
              "nullable": true,
              "type": "string"
            }
          },
          {
            "name": "end",
            "required": false,
            "in": "query",
            "description": "parse by new Date()",
            "schema": {
              "format": "date-time",
              "nullable": true,
              "type": "string"
            }
          },
          {
            "name": "ISBN",
            "required": false,
            "in": "query",
            "schema": {
              "nullable": true,
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "title": "PaginatedBook",
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/Paged"
                    },
                    {
                      "properties": {
                        "data": {
                          "type": "array",
                          "items": {
                            "$ref": "#/components/schemas/Book"
                          }
                        }
                      }
                    }
                  ],
                  "required": [
                    "data",
                    "pageInfo"
                  ]
                }
              }
            }
          }
        },
        "tags": [
          "book"
        ]
      }
    },
    "/api/v1/comment": {
      "post": {
        "operationId": "CommentController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCommentDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "comment"
        ]
      },
      "get": {
        "operationId": "CommentController_getList",
        "parameters": [
          {
            "name": "keyword",
            "required": false,
            "in": "query",
            "schema": {
              "nullable": true,
              "type": "string"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "nullable": true,
              "default": 1,
              "type": "number"
            }
          },
          {
            "name": "size",
            "required": false,
            "in": "query",
            "description": "page size",
            "schema": {
              "nullable": true,
              "default": 10,
              "type": "number"
            }
          },
          {
            "name": "start",
            "required": false,
            "in": "query",
            "description": "parse by new Date()",
            "schema": {
              "format": "date-time",
              "nullable": true,
              "type": "string"
            }
          },
          {
            "name": "end",
            "required": false,
            "in": "query",
            "description": "parse by new Date()",
            "schema": {
              "format": "date-time",
              "nullable": true,
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "comment"
        ]
      }
    },
    "/api/v1/comment/all": {
      "get": {
        "operationId": "CommentController_findAll",
        "parameters": [
          {
            "name": "bookId",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "default": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Comment"
                }
              }
            }
          }
        },
        "tags": [
          "comment"
        ]
      }
    },
    "/api/v1/comment/{id}": {
      "get": {
        "operationId": "CommentController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "comment"
        ]
      },
      "patch": {
        "operationId": "CommentController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateCommentDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "comment"
        ]
      },
      "delete": {
        "operationId": "CommentController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "comment"
        ]
      }
    }
  },
  "info": {
    "title": "examples-books-service",
    "description": "The Books API description",
    "version": "1.0",
    "contact": {}
  },
  "tags": [
    {
      "name": "books",
      "description": ""
    }
  ],
  "servers": [],
  "components": {
    "schemas": {
      "PageInfo": {
        "type": "object",
        "properties": {
          "total": {
            "type": "number"
          }
        },
        "required": [
          "total"
        ]
      },
      "Paged": {
        "type": "object",
        "properties": {
          "pageInfo": {
            "$ref": "#/components/schemas/PageInfo"
          }
        },
        "required": [
          "pageInfo"
        ]
      },
      "Book": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "ISBN": {
            "type": "string"
          },
          "img": {
            "type": "string"
          },
          "description": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "createdAt",
          "updatedAt",
          "name",
          "ISBN",
          "img",
          "description"
        ]
      },
      "CreateCommentDto": {
        "type": "object",
        "properties": {
          "bookId": {
            "type": "string"
          },
          "content": {
            "type": "string"
          },
          "byUserId": {
            "type": "string"
          }
        },
        "required": [
          "bookId",
          "content",
          "byUserId"
        ]
      },
      "Comment": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          },
          "bookId": {
            "type": "string"
          },
          "content": {
            "type": "string"
          },
          "byUserId": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "createdAt",
          "updatedAt",
          "bookId",
          "content",
          "byUserId"
        ]
      },
      "UpdateCommentDto": {
        "type": "object",
        "properties": {
          "bookId": {
            "type": "string"
          },
          "content": {
            "type": "string"
          },
          "byUserId": {
            "type": "string"
          }
        }
      }
    }
  }
}